Issue 38/pydantic 2 (#67)
Added support for `SQLmodel` version `0.0.14` and therefor for `Pydantic 2.0`

Fixes #38

+++ src/my_data/my_data.py
@@ -107,9 +107,10 @@ class MyData:
         """
         self.create_engine()
 
-        if drop_tables:
-            SQLModel.metadata.drop_all(self.database_engine)
-        SQLModel.metadata.create_all(self.database_engine)
+        if self.database_engine:
+            if drop_tables:
+                SQLModel.metadata.drop_all(self.database_engine)
+            SQLModel.metadata.create_all(self.database_engine)
 
     def get_context(self, user: User) -> Context:
         """Get a Context object for this database.
